(Possibly) fix the state management
This uncomments all the tests and lets me observe the behavior I'm
after. This is achieved by letting the CPU spend the first cycle after
reset being idle. This may still be the wrong thing to do, but for now
I'll pretend that this is how it's supposed to work, and progress
further.

Another thing that I'm unsure yet is the fact that my fetch_source and
address registers now lag behind by one cycle. This might be the correct
thing - one cycle changes the registers, in the next they become visible
on posedge(clk), but again, this is something to keep in mind and needs
confirmation.