First attempt to chisel out a register file
This seems to work, but I'm unable to verify the contents of the `regs`
array in GTKWave, so can't be certain. More instructions need to be
implemented to make sure I can read back the same values I put in.

Wiring of `register_file` to the main CPU is still very much WIP.